+++ src/core/Table/TableRowMemoized.tsx
@@ -23,6 +23,7 @@ const TableRow = <T extends Record<string, unknown>>(props: {
   intersectionMargin: number;
   state: TableState<T>; // Needed for explicitly checking selection changes
   onClick?: (event: React.MouseEvent, row: Row<T>) => void;
+  subComponent?: (row: Row<T>) => React.ReactNode;
 }) => {
   const {
     row,
@@ -32,6 +33,7 @@ const TableRow = <T extends Record<string, unknown>>(props: {
     onBottomReached,
     intersectionMargin,
     onClick,
+    subComponent,
   } = props;
 
   const onIntersect = React.useCallback(() => {
@@ -44,26 +46,31 @@ const TableRow = <T extends Record<string, unknown>>(props: {
   });
 
   return (
-    <div
-      {...rowProps}
-      key={rowProps.key}
-      ref={rowRef}
-      onClick={(event) => {
-        onClick?.(event, row);
-      }}
-    >
-      {row.cells.map((cell) => {
-        const cellProps = cell.getCellProps({
-          className: cx('iui-cell', cell.column.cellClassName),
-          style: getCellStyle(cell.column),
-        });
-        return (
-          <div {...cellProps} key={cellProps.key}>
-            {cell.render('Cell')}
-          </div>
-        );
-      })}
-    </div>
+    <>
+      <div
+        {...rowProps}
+        key={rowProps.key}
+        ref={rowRef}
+        onClick={(event) => {
+          onClick?.(event, row);
+        }}
+      >
+        {row.cells.map((cell) => {
+          const cellProps = cell.getCellProps({
+            className: cx('iui-cell', cell.column.cellClassName),
+            style: getCellStyle(cell.column),
+          });
+          return (
+            <div {...cellProps} key={cellProps.key}>
+              {cell.render('Cell')}
+            </div>
+          );
+        })}
+      </div>
+      {row.isExpanded && subComponent && (
+        <div className='iui-row iui-expanded-content'>{subComponent(row)}</div>
+      )}
+    </>
   );
 };
 
@@ -77,6 +84,9 @@ export const TableRowMemoized = React.memo(
     prevProp.row.original === nextProp.row.original &&
     prevProp.state.selectedRowIds?.[prevProp.row.id] ===
       nextProp.state.selectedRowIds?.[nextProp.row.id] &&
+    prevProp.state.expanded?.[prevProp.row.id] ===
+      nextProp.state.expanded?.[nextProp.row.id] &&
+    prevProp.subComponent === nextProp.subComponent &&
     prevProp.row.cells.every(
       (cell, index) => nextProp.row.cells[index].column === cell.column,
     ),