+++ stories/core/Table.stories.tsx
@@ -3,10 +3,21 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import React, { useCallback } from 'react';
-import { CellProps, Column, TableState } from 'react-table';
 import {
+  actions,
+  ActionType,
+  CellProps,
+  Column,
+  Row,
+  TableInstance,
+  TableState,
+} from 'react-table';
+import {
+  Checkbox,
   Code,
+  InputGroup,
   Table,
+  Leading,
   tableFilters,
   TableFilterValue,
   TableProps,
@@ -14,6 +25,7 @@ import {
 import { Story, Meta } from '@storybook/react';
 import { useMemo, useState } from '@storybook/addons';
 import { action } from '@storybook/addon-actions';
+import { CreeveyStoryParams } from 'creevey';
 
 export default {
   title: 'Core/Table',
@@ -319,7 +331,8 @@ export const Filters: Story<TableProps> = (args) => {
     name: string;
     description: string;
     ids: number[];
-    date: Date;
+    startDate: Date;
+    endDate: string;
   };
 
   const translatedLabels = useMemo(
@@ -350,7 +363,7 @@ export const Filters: Story<TableProps> = (args) => {
   );
 
   const tableColumns = useMemo(
-    (): Column[] => [
+    (): Column<TableStoryDataType>[] => [
       {
         Header: 'Table',
         columns: [
@@ -389,11 +402,24 @@ export const Filters: Story<TableProps> = (args) => {
             filter: 'includes',
           },
           {
-            id: 'date',
-            Header: 'Date',
-            accessor: 'date',
+            id: 'startDate',
+            Header: 'Start date',
+            accessor: 'startDate',
             Cell: (props: CellProps<TableStoryDataType>) => {
-              return formatDate(props.row.original.date);
+              return formatDate(props.row.original.startDate);
+            },
+            Filter: tableFilters.DateRangeFilter({
+              translatedLabels,
+            }),
+            filter: 'betweenDate',
+          },
+          {
+            id: 'endDate',
+            Header: 'End date',
+            // Converting string to Date for filtering
+            accessor: (rowData) => new Date(rowData.endDate),
+            Cell: (props: CellProps<TableStoryDataType>) => {
+              return formatDate(new Date(props.row.original.endDate));
             },
             Filter: tableFilters.DateRangeFilter({
               translatedLabels,
@@ -408,14 +434,30 @@ export const Filters: Story<TableProps> = (args) => {
 
   const tableData = useMemo(
     () => [
-      { index: 1, name: 'Name1', description: 'Description1', ids: ['1'] },
+      {
+        index: 1,
+        name: 'Name1',
+        description: 'Description1',
+        ids: ['1'],
+        startDate: new Date('May 1, 2021'),
+        endDate: '2021-05-31T21:00:00.000Z',
+      },
       {
         index: 2,
         name: 'Name2',
         description: 'Description2',
         ids: ['2', '3', '4'],
+        startDate: new Date('May 2, 2021'),
+        endDate: '2021-06-01T21:00:00.000Z',
       },
-      { index: 3, name: 'Name3', description: 'Description3', ids: ['3', '4'] },
+      {
+        index: 3,
+        name: 'Name3',
+        description: 'Description3',
+        ids: ['3', '4'],
+        startDate: new Date('May 3, 2021'),
+        endDate: '2021-06-02T21:00:00.000Z',
+      },
     ],
     [],
   );
@@ -449,24 +491,126 @@ Filters.args = {
       name: 'Name1',
       description: 'Description1',
       ids: ['1'],
-      date: new Date('May 1, 2021'),
+      startDate: new Date('May 1, 2021'),
+      endDate: '2021-05-31T21:00:00.000Z',
     },
     {
       index: 2,
       name: 'Name2',
       description: 'Description2',
       ids: ['2', '3', '4'],
-      date: new Date('May 2, 2021'),
+      startDate: new Date('May 2, 2021'),
+      endDate: '2021-06-01T21:00:00.000Z',
     },
     {
       index: 3,
       name: 'Name3',
       description: 'Description3',
       ids: ['3', '4'],
-      date: new Date('May 3, 2021'),
+      startDate: new Date('May 3, 2021'),
+      endDate: '2021-06-02T21:00:00.000Z',
     },
   ],
   emptyFilteredTableContent: 'No results found. Clear or try another filter.',
+};
+
+export const Expandable: Story<TableProps> = (args) => {
+  const { columns, data, ...rest } = args;
+
+  const onExpand = useCallback(
+    (rows, state) =>
+      action(
+        `Expanded rows: ${JSON.stringify(rows)}. Table state: ${JSON.stringify(
+          state,
+        )}`,
+      )(),
+    [],
+  );
+
+  const tableColumns = useMemo(
+    () => [
+      {
+        Header: 'Table',
+        columns: [
+          {
+            id: 'name',
+            Header: 'Name',
+            accessor: 'name',
+          },
+          {
+            id: 'description',
+            Header: 'Description',
+            accessor: 'description',
+            maxWidth: 200,
+          },
+        ],
+      },
+    ],
+    [],
+  );
+
+  const tableData = useMemo(
+    () => [
+      { name: 'Name1', description: 'Description1' },
+      { name: 'Name2', description: 'Description2' },
+      { name: 'Name3', description: 'Description3' },
+    ],
+    [],
+  );
+
+  const expandedSubComponent = useCallback(
+    (row: Row) => (
+      <div style={{ padding: 16 }}>
+        <Leading>Extra information</Leading>
+        <pre>
+          <code>{JSON.stringify({ values: row.values }, null, 2)}</code>
+        </pre>
+      </div>
+    ),
+    [],
+  );
+
+  return (
+    <Table
+      columns={columns || tableColumns}
+      data={data || tableData}
+      emptyTableContent='No data.'
+      subComponent={expandedSubComponent}
+      onExpand={onExpand}
+      {...rest}
+    />
+  );
+};
+
+Expandable.args = {
+  data: [
+    { name: 'Name1', description: 'Description1' },
+    { name: 'Name2', description: 'Description2' },
+    { name: 'Name3', description: 'Description3' },
+  ],
+  isSelectable: true,
+};
+
+Expandable.parameters = {
+  creevey: {
+    tests: {
+      async open() {
+        const closed = await this.takeScreenshot();
+        const table = await this.browser.findElement({
+          css: '.iui-table',
+        });
+
+        const expanderButtons = await table.findElements({
+          css: '.iui-button',
+        });
+        await expanderButtons[0].click();
+        await expanderButtons[2].click();
+
+        const expanded = await this.takeScreenshot();
+        await this.expect({ closed, expanded }).to.matchImages();
+      },
+    },
+  } as CreeveyStoryParams,
 };
 
 export const LazyLoading: Story<TableProps> = (args) => {
@@ -718,3 +862,127 @@ export const NoData: Story<TableProps> = ({ columns, data, ...rest }) => {
 NoData.args = {
   data: [],
 };
+
+export const ControlledState: Story<TableProps> = (args) => {
+  const { columns, data, ...rest } = args;
+
+  const [selectedRows, setSelectedRows] = useState<Record<string, boolean>>({});
+
+  const tableColumns = useMemo(
+    () => [
+      {
+        Header: 'Table',
+        columns: [
+          {
+            id: 'name',
+            Header: 'Name',
+            accessor: 'name',
+          },
+          {
+            id: 'description',
+            Header: 'Description',
+            accessor: 'description',
+          },
+        ],
+      },
+    ],
+    [],
+  );
+
+  const tableData = useMemo(
+    () => [
+      { name: 'Name1', description: 'Description1' },
+      { name: 'Name2', description: 'Description2' },
+      { name: 'Name3', description: 'Description3' },
+    ],
+    [],
+  );
+
+  const controlledState = useCallback(
+    (state) => {
+      return {
+        ...state,
+        selectedRowIds: { ...selectedRows },
+      };
+    },
+    [selectedRows],
+  );
+
+  // When using `useControlledState` we are fully responsible for the state part we are modifying.
+  // Therefore we want to keep our outside state (`selectedRows`) in sync with inside table state (`state.selectedRowIds`).
+  const tableStateReducer = (
+    newState: TableState,
+    action: ActionType,
+    previousState: TableState,
+    instance?: TableInstance,
+  ): TableState => {
+    switch (action.type) {
+      case actions.toggleRowSelected: {
+        const newSelectedRows = {
+          ...selectedRows,
+        };
+        if (action.value) {
+          newSelectedRows[action.id] = true;
+        } else {
+          delete newSelectedRows[action.id];
+        }
+        setSelectedRows(newSelectedRows);
+        newState.selectedRowIds = newSelectedRows;
+        break;
+      }
+      case actions.toggleAllRowsSelected: {
+        if (!instance?.rowsById) {
+          break;
+        }
+        const newSelectedRows = {} as Record<string, boolean>;
+        if (action.value) {
+          Object.keys(instance.rowsById).forEach(
+            (id) => (newSelectedRows[id] = true),
+          );
+        }
+        setSelectedRows(newSelectedRows);
+        newState.selectedRowIds = newSelectedRows;
+        break;
+      }
+      default:
+        break;
+    }
+    return newState;
+  };
+
+  return (
+    <>
+      <InputGroup label='Control selected rows' style={{ marginBottom: 11 }}>
+        {tableData.map((data, index) => (
+          <Checkbox
+            key={index}
+            label={data.name}
+            checked={selectedRows[index]}
+            onChange={(e) => {
+              setSelectedRows((rowIds) => {
+                const selectedRowIds = { ...rowIds };
+                if (e.target.checked) {
+                  selectedRowIds[index] = true;
+                } else {
+                  delete selectedRowIds[index];
+                }
+                return selectedRowIds;
+              });
+            }}
+          />
+        ))}
+      </InputGroup>
+      <Table
+        columns={columns || tableColumns}
+        data={data || tableData}
+        emptyTableContent='No data.'
+        useControlledState={controlledState}
+        stateReducer={tableStateReducer}
+        isSelectable
+        {...rest}
+      />
+    </>
+  );
+};
+
+ControlledState.args = { isSelectable: true };